etcd: Use IPv6 IP addresses
Update the etcd-member template and the environment input into the
template to support an IPv6 control plane.

Localhost and "all addresses" can't be hard coded, as it's different
for IPv4 vs IPv6.

When using an IPv6 address in a URL, we need to wrap it in square
brackets.  This allows a suffix of ":port" to still be parsed
correctly.

I left the variable name IPV4_ADDRESS in place to cut down on the
required changes at this point, but the name should probably be
changed at some point.